hxgn-sdx-bot/bot/ocr-engine/requirements.txt
playwright
google-genai
python-dotenv
pymupdf
python-dateutil
requests
